loginDao update

+++ src/main/java/com/springrest/springrest/dao/LoginDao.java
@@ -6,7 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.springrest.springrest.entities.Login;
 
-public interface LoginDao extends JpaRepository<Login, Long> {
-
+public interface LoginDao extends JpaRepository<Login, Integer> {
     Optional<Login> findByName(String name);
 }
